detailUI
detail page navigation to view
